fix(platform): make settings the only preferences surface
Route legacy preference links into unified Settings, place Morning Brief beside time zone, and remove the retired Preferences page.
