+++ turbo/apps/platform/src/i18n/locales/en-US/common.json
@@ -5089,7 +5089,6 @@
           "upstreamModel": "Upstream model"
         },
         "capture": {
-          "description": "Capture HTTP header names, selected safe header values, and bodies in network logs for debugging.",
           "disabled": "Disabled",
           "enabled_one": "Enabled for the next {{count}} run",
           "enabled_other": "Enabled for the next {{count}} runs",
@@ -5112,10 +5111,8 @@
           "recovery": "recovery",
           "title": "Realtime connection diagnostics",
           "unavailable": "unavailable"
-        },
-        "tab": "Debug"
+        }
       },
-      "documentTitle": "Preferences",
       "language": {
         "description": "Choose your preferred language for the {{brandName}} interface",
         "label": "Language",
@@ -5145,8 +5142,6 @@
         "retryMessage": "Morning Brief settings could not be loaded or updated. Retry to continue.",
         "title": "Morning Brief"
       },
-      "pageDescription": "Manage your appearance and agent runtime preferences",
-      "pageTitle": "Preferences",
       "send": {
         "cmdEnter": "⌘ Enter",
         "cmdEnterDescription": "Press ⌘/Ctrl+Enter to send, Enter for new line",
@@ -5155,11 +5150,6 @@
         "enterDescription": "Press Enter to send, Shift+Enter for new line",
         "sectionTitle": "Enter",
         "title": "Send message with"
-      },
-      "tabs": {
-        "appearance": "Appearance",
-        "models": "Personal Models",
-        "timezone": "Time Zone"
       },
       "timezone": {
         "description": "Times shown to you and used for automation runs.",
