+++ turbo/apps/platform/src/i18n/locales/hi-IN/common.json
@@ -5097,7 +5097,6 @@
           "upstreamModel": "अपस्ट्रीम मॉडल"
         },
         "capture": {
-          "description": "डिबगिंग के लिए नेटवर्क लॉग में HTTP हेडर नाम, चयनित सुरक्षित हेडर मान और बॉडी कैप्चर करें।",
           "disabled": "अक्षम",
           "enabled_one": "अगले {{count}} रन के लिए सक्षम",
           "enabled_other": "अगले {{count}} रन के लिए सक्षम",
@@ -5120,10 +5119,8 @@
           "recovery": "पुनर्प्राप्ति",
           "title": "रीयल-टाइम कनेक्शन निदान",
           "unavailable": "अनुपलब्ध"
-        },
-        "tab": "डिबग"
+        }
       },
-      "documentTitle": "प्राथमिकताएँ",
       "language": {
         "description": "{{brandName}} इंटरफ़ेस के लिए अपनी पसंदीदा भाषा चुनें",
         "label": "भाषा",
@@ -5153,8 +5150,6 @@
         "retryMessage": "Morning Brief सेटिंग लोड या अपडेट नहीं हो सकीं। जारी रखने के लिए फिर से कोशिश करें।",
         "title": "सुबह की संक्षिप्त जानकारी"
       },
-      "pageDescription": "अपनी उपस्थिति और एजेंट रनटाइम प्राथमिकताएँ प्रबंधित करें।",
-      "pageTitle": "प्राथमिकताएँ",
       "send": {
         "cmdEnter": "⌘ दर्ज करें",
         "cmdEnterDescription": "भेजने के लिए ⌘/Ctrl+Enter दबाएँ, नई लाइन के लिए Enter दबाएँ",
@@ -5163,11 +5158,6 @@
         "enterDescription": "भेजने के लिए Enter दबाएँ, नई लाइन के लिए Shift+Enter दबाएँ",
         "sectionTitle": "एंटर",
         "title": "के साथ संदेश भेजें"
-      },
-      "tabs": {
-        "appearance": "उपस्थिति",
-        "models": "व्यक्तिगत मॉडल",
-        "timezone": "समय क्षेत्र"
       },
       "timezone": {
         "description": "समय आपको दिखाया गया है और ऑटोमेशन चलाने के लिए उपयोग किया जाता है।",
