+++ turbo/apps/platform/src/i18n/locales/it-IT/common.json
@@ -5155,7 +5155,6 @@
           "upstreamModel": "Modello upstream"
         },
         "capture": {
-          "description": "Cattura i nomi delle intestazioni HTTP, i valori delle intestazioni sicure selezionate e i corpi nei log di rete per il debug.",
           "disabled": "Disabilitato",
           "enabled_one": "Abilitato per la prossima esecuzione {{count}}",
           "enabled_many": "Abilitato per le prossime {{count}} esecuzioni",
@@ -5179,10 +5178,8 @@
           "recovery": "recupero",
           "title": "Diagnostica della connessione in tempo reale",
           "unavailable": "non disponibile"
-        },
-        "tab": "Debug"
+        }
       },
-      "documentTitle": "Preferenze",
       "language": {
         "description": "Scegli la tua lingua preferita per l'interfaccia {{brandName}}",
         "label": "Lingua",
@@ -5212,8 +5209,6 @@
         "retryMessage": "Non è stato possibile caricare o aggiornare le impostazioni di Morning Brief. Riprova per continuare.",
         "title": "Riepilogo mattutino"
       },
-      "pageDescription": "Gestisci l'aspetto e le preferenze di esecuzione degli agenti",
-      "pageTitle": "Preferenze",
       "send": {
         "cmdEnter": "⌘ Invio",
         "cmdEnterDescription": "Premi ⌘/Ctrl+Invio per inviare, Invio per una nuova riga",
@@ -5222,11 +5217,6 @@
         "enterDescription": "Premi Invio per inviare, Maiusc+Invio per una nuova riga",
         "sectionTitle": "Invio",
         "title": "Invia messaggio con"
-      },
-      "tabs": {
-        "appearance": "Aspetto",
-        "models": "Modelli personali",
-        "timezone": "Fuso orario"
       },
       "timezone": {
         "description": "Orari mostrati e usati per le esecuzioni delle automazioni.",
