+++ turbo/apps/platform/src/i18n/locales/ko-KR/common.json
@@ -5089,7 +5089,6 @@
           "upstreamModel": "업스트림 모델"
         },
         "capture": {
-          "description": "디버깅을 위해 HTTP 헤더 이름, 선택된 안전 헤더 값 및 본문을 네트워크 로그에 캡처합니다.",
           "disabled": "비활성화됨",
           "enabled_one": "다음 {{count}}회 실행 동안 활성화됨",
           "enabled_other": "다음 {{count}}회 실행 동안 활성화됨",
@@ -5112,10 +5111,8 @@
           "recovery": "복구",
           "title": "실시간 연결 진단",
           "unavailable": "사용 불가"
-        },
-        "tab": "디버그"
+        }
       },
-      "documentTitle": "환경설정",
       "language": {
         "description": "{{brandName}} 인터페이스에 사용할 기본 언어를 선택하세요",
         "label": "언어",
@@ -5145,8 +5142,6 @@
         "retryMessage": "Morning Brief 설정을 불러오거나 업데이트할 수 없습니다. 계속하려면 다시 시도하세요.",
         "title": "모닝 브리프"
       },
-      "pageDescription": "외관 및 에이전트 실행 환경설정을 관리하세요",
-      "pageTitle": "환경설정",
       "send": {
         "cmdEnter": "⌘ Enter",
         "cmdEnterDescription": "⌘/Ctrl+Enter를 눌러 전송, Enter는 새 줄",
@@ -5155,11 +5150,6 @@
         "enterDescription": "Enter를 눌러 전송, Shift+Enter는 새 줄",
         "sectionTitle": "Enter",
         "title": "메시지 전송 키"
-      },
-      "tabs": {
-        "appearance": "외관",
-        "models": "개인 모델",
-        "timezone": "시간대"
       },
       "timezone": {
         "description": "표시되는 시간과 자동화 실행에 사용되는 시간입니다.",
