+++ turbo/apps/platform/src/signals/okou-page/settings/send-mode-preference.ts
@@ -1,0 +1,34 @@
+import { command, computed, state } from "ccstate";
+import type { SendMode } from "@okouai/api-contracts/contracts/user-preferences";
+
+import { sendMode$ } from "../../send-mode.ts";
+import { updateUserPreference$ } from "./user-preferences.ts";
+
+/**
+ * Tracks the send mode value most recently submitted via updateSendMode$.
+ * Used by the view to show an optimistic spinner on the correct button.
+ * Cleared automatically when the command completes or fails.
+ */
+const internalPendingSendMode$ = state<SendMode | null>(null);
+
+export const pendingSendMode$ = computed((get) => {
+  return get(internalPendingSendMode$);
+});
+
+/**
+ * Update send mode preference. After saving, await the refetched value so the
+ * UI never flashes back to the old value before the signal updates.
+ */
+export const updateSendMode$ = command(
+  async ({ get, set }, value: SendMode, signal: AbortSignal) => {
+    set(internalPendingSendMode$, value);
+    await set(updateUserPreference$, { sendMode: value }, signal).finally(
+      () => {
+        set(internalPendingSendMode$, null);
+      },
+    );
+    signal.throwIfAborted();
+    await get(sendMode$);
+    signal.throwIfAborted();
+  },
+);
