+++ turbo/apps/platform/src/views/okou-page/components/settings/sections/preference-section.tsx
@@ -16,8 +16,9 @@ import { detach, Reason } from "../../../../../signals/utils.ts";
 import {
   updateSendMode$,
   pendingSendMode$,
-} from "../../../../../signals/okou-page/settings/preferences-page.ts";
+} from "../../../../../signals/okou-page/settings/send-mode-preference.ts";
 import { TimezoneSettings } from "../timezone-settings.tsx";
+import { MorningBriefSettings } from "../morning-brief-settings.tsx";
 import { SettingsSectionHeading } from "../settings-section-heading.tsx";
 import { AccountSection } from "./account-section.tsx";
 import { LanguageSettings } from "../language-settings.tsx";
@@ -243,6 +244,7 @@ export function PreferenceSection() {
           })}
         />
         <TimezoneSettings />
+        <MorningBriefSettings />
       </section>
     </div>
   );
